+++ ceammc/ext/src/list/list_walk.cpp
@@ -5,10 +5,10 @@
 using namespace ceammc;
 using namespace ceammc::list;
 
-static const char* SYM_SINGLE = "single";
-static const char* SYM_CLIP = "clip";
-static const char* SYM_WRAP = "wrap";
-static const char* SYM_FOLD = "fold";
+static t_symbol* SYM_SINGLE;
+static t_symbol* SYM_CLIP;
+static t_symbol* SYM_WRAP;
+static t_symbol* SYM_FOLD;
 
 ListWalk::ListWalk(const PdArgs& a)
     : BaseObject(a)
@@ -19,23 +19,24 @@ ListWalk::ListWalk(const PdArgs& a)
     , single_done_(false)
 {
     createOutlet();
+    createOutlet();
 
     createProperty(new PointerProperty<bool>("@direction", &forward_, false));
     createProperty(new PointerProperty<int>("@length", &length_, false));
     createProperty(new PointerProperty<AtomList>("@value", &lst_));
 
-    walk_mode_ = new SymbolEnumProperty("@mode", gensym(SYM_SINGLE));
-    walk_mode_->appendEnum(gensym(SYM_WRAP));
-    walk_mode_->appendEnum(gensym(SYM_CLIP));
-    walk_mode_->appendEnum(gensym(SYM_FOLD));
+    walk_mode_ = new SymbolEnumProperty("@mode", SYM_SINGLE);
+    walk_mode_->appendEnum(SYM_WRAP);
+    walk_mode_->appendEnum(SYM_CLIP);
+    walk_mode_->appendEnum(SYM_FOLD);
     createProperty(walk_mode_);
 
     // aliases
-    createProperty(new SymbolEnumAlias("@single", walk_mode_, gensym(SYM_SINGLE)));
-    createProperty(new SymbolEnumAlias("@loop", walk_mode_, gensym(SYM_WRAP)));
-    createProperty(new SymbolEnumAlias("@wrap", walk_mode_, gensym(SYM_WRAP)));
-    createProperty(new SymbolEnumAlias("@clip", walk_mode_, gensym(SYM_CLIP)));
-    createProperty(new SymbolEnumAlias("@fold", walk_mode_, gensym(SYM_FOLD)));
+    createProperty(new SymbolEnumAlias("@single", walk_mode_, SYM_SINGLE));
+    createProperty(new SymbolEnumAlias("@loop", walk_mode_, SYM_WRAP));
+    createProperty(new SymbolEnumAlias("@wrap", walk_mode_, SYM_WRAP));
+    createProperty(new SymbolEnumAlias("@clip", walk_mode_, SYM_CLIP));
+    createProperty(new SymbolEnumAlias("@fold", walk_mode_, SYM_FOLD));
 
     createCbProperty("@size", &ListWalk::p_size);
     createCbProperty("@index", &ListWalk::p_index, &ListWalk::p_set_index);
@@ -81,7 +82,7 @@ AtomList ListWalk::p_size() const { return AtomList(lst_.size()); }
 
 AtomList ListWalk::p_index() const
 {
-    if (walk_mode_->value() != gensym(SYM_FOLD))
+    if (walk_mode_->value() != SYM_FOLD)
         return AtomList(current_pos_);
     else {
         size_t idx = 0;
@@ -128,7 +129,7 @@ void ListWalk::toPosition(int pos)
 
     size_t idx = 0;
 
-    if (walk_mode_->value() == gensym(SYM_SINGLE)) {
+    if (walk_mode_->value() == SYM_SINGLE) {
         if (pos < 0) {
             current_pos_ = 0;
             single_done_ = true;
@@ -139,13 +140,13 @@ void ListWalk::toPosition(int pos)
             current_pos_ = lst_.size() - 1;
             single_done_ = true;
         }
-    } else if (walk_mode_->value() == gensym(SYM_WRAP)) {
+    } else if (walk_mode_->value() == SYM_WRAP) {
         if (calcWrapIndex(pos, lst_.size(), &idx))
             current_pos_ = idx;
-    } else if (walk_mode_->value() == gensym(SYM_CLIP)) {
+    } else if (walk_mode_->value() == SYM_CLIP) {
         if (calcClipIndex(pos, lst_.size(), &idx))
             current_pos_ = idx;
-    } else if (walk_mode_->value() == gensym(SYM_FOLD)) {
+    } else if (walk_mode_->value() == SYM_FOLD) {
         current_pos_ = pos;
     } else {
         OBJ_ERR << "unsupported list mode: " << walk_mode_->value()->s_name;
@@ -165,28 +166,37 @@ void ListWalk::current()
     }
 
     //! single
-    if (walk_mode_->value() == gensym(SYM_SINGLE)) {
+    if (walk_mode_->value() == SYM_SINGLE) {
         if (single_done_)
             return;
 
         listTo(0, lst_.slice(current_pos_, current_pos_ + length_ - 1));
     }
     //! clip
-    else if (walk_mode_->value() == gensym(SYM_CLIP)) {
+    else if (walk_mode_->value() == SYM_CLIP) {
         listTo(0, list::sliceClip(lst_, current_pos_, length_));
     }
     //! wrap/loop
-    else if (walk_mode_->value() == gensym(SYM_WRAP)) {
+    else if (walk_mode_->value() == SYM_WRAP) {
         listTo(0, list::sliceWrap(lst_, current_pos_, length_));
     }
     //! fold
-    else if (walk_mode_->value() == gensym(SYM_FOLD)) {
+    else if (walk_mode_->value() == SYM_FOLD) {
         listTo(0, list::sliceFold(lst_, current_pos_, length_));
     }
+
+    // last element
+    if ((current_pos_ + 1) == lst_.size())
+        bangTo(1);
 }
 
 extern "C" void setup_list0x2ewalk()
 {
+    SYM_SINGLE = gensym("single");
+    SYM_CLIP = gensym("clip");
+    SYM_WRAP = gensym("wrap");
+    SYM_FOLD = gensym("fold");
+
     ObjectFactory<ListWalk> obj("list.walk");
     obj.addMethod("current", &ListWalk::m_current);
     obj.addMethod("next", &ListWalk::m_next);